+++ Scripts/Utils.gd
@@ -1,4 +1,4 @@
-enum TileColor{WHITE, BLACK, BLUE, RED, PINK}
+enum TileColor{WHITE, BLACK, BLUE, RED, PINK, GREY, GREEN}
 
 static func flat(arr):
 	var flat_list = []
@@ -15,3 +15,11 @@ static func shuffle(arr):
 	for i in range(arr.size()):
 		sampled.append( shuffled.pop_front() )
 	return sampled
+
+static func fixScale(imageSize, viewportSize):
+	var scaleFactor = 1
+	if viewportSize.aspect() > imageSize.aspect():
+		scaleFactor = viewportSize.y/imageSize.y
+	else:
+		scaleFactor = viewportSize.x/imageSize.x
+	return Vector2(scaleFactor,scaleFactor)